Build API for user responses

+++ src/server/api/routers/form.ts
@@ -1,9 +1,12 @@
 import { z } from "zod"
 
-import { createTRPCRouter, publicProcedure } from "~/server/api/trpc"
+import {
+  createTRPCRouter,
+  protectedProcedure,
+  publicProcedure,
+} from "~/server/api/trpc"
 
 // Validation schemas
-
 const formBlockSchema = z.object({
   order: z.number(),
   title: z.string().optional(),
@@ -27,7 +30,8 @@ const formBlockSchema = z.object({
 })
 
 const formResponsesSchema = z.object({
-  answers: z.any(),
+  formId: z.string(),
+  answers: z.object({}).passthrough(),
 })
 
 const formSchema = z.object({
@@ -44,37 +48,39 @@ const formSchema = z.object({
 })
 
 // Router
-
 export const formRouter = createTRPCRouter({
-  getUserForms: publicProcedure
+  // Form procedures
+  getFormById: protectedProcedure
     .input(z.string())
     .query(async ({ ctx, input }) => {
-      const forms = ctx.db.user.findUnique({
+      const form = await ctx.db.form.findUnique({
         where: {
-          clerkUserId: input,
-        },
-        select: {
-          forms: true,
+          id: input,
+          userId: ctx.authenticatedUser.userId,
         },
       })
-      return forms
+      if (!form) {
+        throw new Error("Form could not be found or you were denied access")
+      }
+
+      return form
     }),
-  getFormById: publicProcedure
+  getPublicFormById: publicProcedure
     .input(z.string())
     .query(async ({ ctx, input }) => {
-      const form = ctx.db.form.findUnique({
+      const form = await ctx.db.form.findUnique({
         where: {
           id: input,
         },
       })
       return form
     }),
-  createForm: publicProcedure
+  createForm: protectedProcedure
     .input(formSchema)
     .mutation(async ({ ctx, input }) => {
-      const form = ctx.db.form.create({
+      const form = await ctx.db.form.create({
         data: {
-          userId: input.userId,
+          userId: ctx.authenticatedUser.userId,
           title: input.title,
           startDate: input.startDate,
           endDate: input.endDate,
@@ -90,10 +96,24 @@ export const formRouter = createTRPCRouter({
       return form
     }),
 
-  deleteForm: publicProcedure
+  deleteForm: protectedProcedure
     .input(z.string())
     .mutation(async ({ ctx, input }) => {
       await ctx.db.$transaction(async db => {
+        // First, check if the form belongs to the logged-in user
+        const form = await db.form.findUnique({
+          where: {
+            id: input,
+            userId: ctx.authenticatedUser.userId,
+          },
+        })
+
+        if (!form) {
+          throw new Error(
+            "Form not found or you do not have permission to delete this form"
+          )
+        }
+
         // Delete all associated data before deleting the form
         await db.formBlock.deleteMany({
           where: { formId: input },
@@ -103,12 +123,59 @@ export const formRouter = createTRPCRouter({
         })
 
         // Finally, delete the form
-        const form = db.form.delete({
+        const deletedForm = db.form.delete({
           where: {
             id: input,
           },
         })
-        return form
+        return deletedForm
       })
+    }),
+
+  // Form responses procedures
+  getResponses: protectedProcedure
+    .input(z.string())
+    .query(async ({ ctx, input }) => {
+      const form = await ctx.db.form.findUnique({
+        where: {
+          id: input,
+          userId: ctx.authenticatedUser.userId,
+        },
+      })
+
+      if (!form) {
+        throw new Error(
+          "Form not found or you do not have permission to delete this form"
+        )
+      }
+
+      const responses = await ctx.db.response.findMany({
+        where: {
+          formId: input,
+        },
+      })
+
+      return responses
+    }),
+  addResponse: publicProcedure
+    .input(formResponsesSchema)
+    .mutation(async ({ ctx, input }) => {
+      const response = await ctx.db.response.create({
+        data: {
+          formId: input.formId,
+          answers: input.answers,
+        },
+      })
+      return response
+    }),
+  deleteResponse: protectedProcedure
+    .input(z.string())
+    .mutation(async ({ ctx, input }) => {
+      const response = await ctx.db.response.delete({
+        where: {
+          id: input,
+        },
+      })
+      return response
     }),
 })
